requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
langchain-community
python-dotenv
pydantic
pymupdf
